requirements.txt
Flask
requests
Pillow
ultralytics
torch